Move expr_builder inside expression
