+++ include/autoppl/autoppl.hpp
@@ -9,7 +9,7 @@
 #include "expression/variable/binop.hpp"
 #include "expression/variable/constant.hpp"
 #include "expression/variable/variable_viewer.hpp"
+#include "expression/expr_builder.hpp"
 #include "mcmc/mh.hpp"
 #include "mcmc/hmc/nuts/nuts.hpp"
-#include "expr_builder.hpp"
 #include "variable.hpp"
